+++ src/app/user/components/UserList.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { GetUsersList } from '../../redux/actions/user/GetUsersList';
+import { GetUsersList } from '../../../redux/actions/user/GetUsersList';
 import { Link } from 'react-router-dom'
 import { ListGroup, Button } from 'react-bootstrap'
